Add reader data store to @automattic/data-stores
The new onboarding code doesn't use the main Calypso redux store, so it
doesn't know whether a user is an a11n. Adding reader data to the stores
in `@automattic/data-stores` gives it access to the reader team data,
which is how the client can know whether a user is an a11n.

This information is needed to only show the new font smoothing settings
to a12s.